requirements.txt
gradio
aiohttp
aiosignal
altgraph
async-timeout
attrs
certifi
charset-normalizer
decorator
edge-tts
frozenlist
idna
imageio
imageio-ffmpeg
macholib
moviepy
multidict
numpy
Pillow
proglog
pyinstaller
pyinstaller-hooks-contrib
PyYAML
requests
tqdm
urllib3
yarl

